Add ScriptHandler for symlink creation on install/update
Implemented a more robust way to handle the creation of a symlink for the application on installation or update. Created a new file `ScriptHandler.php` that has a `createSymlink()` method. This method will check if a symlink or file already exists and will abort the process not to overwrite it. Changes are also made in the `composer.json` file, replacing the `post-install-cmd` and `post-update-cmd` scripts from `php setup.php` to `Rover\\ScriptHandler::createSymlink`. This offers a more controlled and error-free process.