Update symlink creation in ScriptHandler
Updated the `createSymlink` function in `ScriptHandler`. The method now considers two possible locations of the `robo` executable - top-level and nested vendor directories. This is to ensure the successful creation of the symbolic link regardless of the installation type of the 'consolidation/robo' package. Also added a check and creation of destination directory if it does not exist.